// frontend/src/openapi/generated/model/cate-id-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cats example
 * The cats API description
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CateIdInput
 */
export interface CateIdInput {
    /**
     * 
     * @type {number}
     * @memberof CateIdInput
     */
    'cateId': number | null;
}