// frontend/src/openapi/generated/model/get-memos-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cats example
 * The cats API description
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GetMemosInput
 */
export interface GetMemosInput {
    /**
     * 
     * @type {string}
     * @memberof GetMemosInput
     */
    'search': string | null;
    /**
     * 
     * @type {number}
     * @memberof GetMemosInput
     */
    'offset': number;
    /**
     * 
     * @type {number}
     * @memberof GetMemosInput
     */
    'limit': number;
    /**
     * 
     * @type {number}
     * @memberof GetMemosInput
     */
    'cateQueryStr': number | null;
    /**
     * 
     * @type {string}
     * @memberof GetMemosInput
     */
    'tagQueryStr': string | null;
    /**
     * 
     * @type {string}
     * @memberof GetMemosInput
     */
    'menuQueryStr': string | null;
}